new file:   artifacts/build-info/8bbef55b8dca557cae6c296b6d71e749.json 	deleted:    artifacts/build-info/b16b16cf1fcb9ebb084d9ac8d5912659.json 	modified:   artifacts/contracts/gasChallenge.sol/gasChallenge.dbg.json 	modified:   artifacts/contracts/gasChallenge.sol/gasChallenge.json 	modified:   cache/solidity-files-cache.json 	modified:   contracts/gasChallenge.sol

+++ contracts/gasChallenge.sol
@@ -7,7 +7,7 @@ contract gasChallenge {
 
     //Function to check for sum of array
     //No changes required in this function
-    function getSumOfArray() public view returns (uint) {
+    function getSumOfArray() public view returns (uint256) {
         uint sum = 0;
         for (uint i = 0; i < numbers.length; i++) {
             sum += numbers[i];
